refactoring, moving Login and Main components
